+++ src/types/conditional.types.ts
@@ -1,0 +1,4 @@
+export interface TimeWindow {
+  start: string
+  end: string
+}